// src/app/pages/Models/DatosRevisorFiscalDto.ts
export interface DatosRevisorFiscalDto {
    id: number;
    idFormulario: number;
    tieneRevisorFiscal: boolean;
    justificarRespuesta: string;
    revisorFiscalAdscritoFirma: boolean;
    nombreFirma: string;
    nombreCompletoApellidos: string;
    tipoID: string;
    numeroID: string;
    telefono: string;
    ciudad: string;
    direccion: string;
    email: string;
  } 
  